feat: add Exercises 4.21.- 4.23.

+++ part4/blog-list/controllers/blogs.js
@@ -1,7 +1,6 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
-const jwt = require('jsonwebtoken')
 
 
 
@@ -13,14 +12,7 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
-
-  const token = request.token
-  const decodedToken = jwt.verify(token, process.env.SECRET)
-  if (!token || !decodedToken.id) {
-    return response.status(401).json({ error: 'token missing or invalid' })
-  }
-
-  const user = await User.findById(decodedToken.id)
+  const user = await User.findById(request.user)
 
   const blog = new Blog({
     title: body.title,
@@ -29,17 +21,34 @@ blogsRouter.post('/', async (request, response) => {
     likes: body.likes,
     user: user._id
   })
+
   const savedBlog = await blog.save()
   user.blogs = user.blogs.concat(savedBlog._id)
   user._id = user._id.toString()
+  console.log(user)
   await user.save()
 
   response.json(savedBlog.toJSON())
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+
+  const blog = await Blog.findById(request.params.id)
+  if (!blog || !blog.user) {
+    return response.status(404).json({ error: 'Not found' })
+  }
+  if (blog.user.toString() === request.user) {
+
+    await Blog.findByIdAndRemove(request.params.id)
+    const user = await User.findById(request.user)
+    user.blogs = user.blogs.filter((blog) => (blog.toString() !== request.params.id))
+    await user.save()
+    response.status(204).end()
+  } else {
+    return response.status(404).json({ error: 'Not found*' })
+  }
+
+
 })
 
 blogsRouter.put('/:id', (request, response, next) => {
